cyber_inquisitor/inquisitor/pyproject.toml
[tool.poetry]
name = "inquisitor"
version = "0.1.0"
description = ""
authors = ["allemago <gbk.maelle@outlook.fr>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pypcap = "^1.3.0"
dpkt = "^1.9.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.scripts]
inquisitor = "src.cli:run"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
